+++ src/Modules/Tours/Explorer.Tours.API/Public/Administration/IEquipmentService.cs
@@ -9,5 +9,5 @@ public interface IEquipmentService
     Result<PagedResult<EquipmentDto>> GetPaged(int page, int pageSize);
     Result<EquipmentDto> Create(EquipmentDto equipment);
     Result<EquipmentDto> Update(EquipmentDto equipment);
-    Result Delete(int id);
+    Result Delete(long id);
 }
